+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Commit processing is moved to server-side
 - Upgrade to Next.js 15.4
 - Upgrade to ESLint 9
+- Upgrade Commitspark libraries
 
 ## [0.30.0] - 2025-03-19